Add files via upload
Have done some more work on flashing colours however is not working properly yet. We have added a new feature to the output window where the window flashes different colours after clicking the flash button. In addition, have started coding the if statements and variables however this feature is not fully functional, after hitting an issue with the code. The program will not work with this commit as the code is not fully complete please look for the next commit for the working code.