Add files via upload
As the main commit was not working, i have created a separate branch for this final commit. This commit includes variables, if statements and flashing colours for part 2 and for part 1 i have completed the unit testing and some tests for part 2. I couldnt upload the full test environment as it is over 100 files so i have uploaded the unittest1.cs file and the programming testenvironment c# project file